Build: Game runs with a while-true loop window

+++ src/config/flecs_data.zig
@@ -446,7 +446,6 @@ pub const StaticMesh = struct {
     materials: [renderer.sub_mesh_max_count]MaterialHandle,
 };
 
-
 pub const HierarchicalStaticMesh = struct {
     static_mesh_count: u32,
     static_meshes: [renderer.mesh_lod_max_count]StaticMesh,
@@ -734,10 +733,10 @@ pub const CompCombatant = struct {
 pub const EnvironmentInfo = struct {
     paused: bool,
     active_camera: ?ecsu.Entity,
-    time_multiplier: f32 = 1.0,
-    world_time: f32,
-    time_of_day_percent: f32,
-    sun_height: f32,
+    time_multiplier: f64 = 1.0,
+    world_time: f64,
+    time_of_day_percent: f64,
+    sun_height: f64,
     sky_light: ?ecsu.Entity,
     sun: ?ecsu.Entity,
     player: ?ecsu.Entity,
